+++ components/Screens/ResetPasswordScreen.js
@@ -3,7 +3,7 @@ import * as firebase from 'firebase';
 
 import { View, Text, TouchableOpacity, StyleSheet, TextInput, Alert } from 'react-native';
 
-import { NanumText } from '../../components/StyledText';
+import { NanumText } from '../StyledText';
 
 const styles = StyleSheet.create({
   container: {
